+++ lib/src/oauth2_utils.dart
@@ -30,8 +30,8 @@ class OAuth2Utils {
       } else {
         val = v.trim();
       }
-      // qsList.add(k + '=' + Uri.encodeComponent(val));
-      qsList.add(k + '=' + val);
+      qsList.add(k + '=' + Uri.encodeComponent(val));
+      // qsList.add(k + '=' + val);
     });
 
     return qsList.join('&');